django/requirements.txt
Django==2.2.4
django-admin-multiple-choice-list-filter==0.1.1
django-extensions==2.2.1
django-sequences==2.2
django-smart-selects==1.5.4
django-werkzeug==1.0.0
lxml==4.3.1
python-barcode==0.9.0
python-docx==0.8.10
python-stdnum==1.10
pytz==2018.5
six==1.11.0
sqlparse==0.3.0
Werkzeug==0.14.1
